requirements.txt
Django==1.4.1
distribute==0.6.28
dj-database-url==0.2.1
django-cors==0.0.1
django-debug-toolbar==0.9.4
django-tastypie==0.9.11
mimeparse==0.1.3
python-dateutil==1.5
